# feat(type-safe-api): enable snapstart for java handlers by default and optimise java interceptors