--- a/packages/type-safe-api/scripts/type-safe-api/generators/java-lambda-handlers/templates/handlers.handlebars
+++ b/packages/type-safe-api/scripts/type-safe-api/generators/java-lambda-handlers/templates/handlers.handlebars
@@ -32,12 +32,17 @@ import java.util.List;
  */
 public class {{operationIdCamelCase}}Handler extends {{operationIdCamelCase}} {
     /**
+     * Interceptors are initialised once during the lambda "init" phase
+     */
+    private final List<Interceptor<{{operationIdCamelCase}}Input>> interceptors = DefaultInterceptors.all();
+
+    /**
      * Return the interceptors for this handler.
      * You can also use the @Interceptors annotation on the class to add interceptors
      */
     @Override
     public List<Interceptor<{{operationIdCamelCase}}Input>> getInterceptors() {
-        return DefaultInterceptors.all();
+        return this.interceptors;
     }
 
     /**