--- a/packages/type-safe-api/src/construct/functions/index.ts
+++ b/packages/type-safe-api/src/construct/functions/index.ts
@@ -1,0 +1,3 @@
+/*! Copyright [Amazon.com](http://amazon.com/), Inc. or its affiliates. All Rights Reserved.
+SPDX-License-Identifier: Apache-2.0 */
+export * from "./snap-start-java-function";