--- a/packages/type-safe-api/src/construct/index.ts
+++ b/packages/type-safe-api/src/construct/index.ts
@@ -5,3 +5,4 @@ export * from "./waf/types";
 export * from "./authorizers";
 export * from "./integrations";
 export * from "./spec";
+export * from "./functions";