Java handlers now have snapstart enabled by default to improve cold start times.

Ensure logging, metrics and tracing interceptors perform as much initialisation as possible during the init phase which can therefore be saved from snapstart invocations.

Also move the annotation processing to the init phase since this takes approximately a second.

This reduces cold start times from ~25s down to ~4s. Warm invocations remain the same at <100ms.

On a cold start it seems as though the logger takes about 500ms per log, and the tracer takes about 1500ms to end the subsegment - so there may be an opportunity to reduce the 4s cold start even further if we can figure out why :)

Fixes #572